refactor the api.js and divide into createTemplate.js, global.js and updateContent.js relates #30
