+++ index.html
@@ -21,15 +21,6 @@
 			</select>
 	</header>
 
-	<!-- <section class="menu">
-		<select>
-  		<option value="Israel">Israel</option>
-  		<option value="UK">UK</option>
-  		<option value="South Korea">South Korea</option>
-  		<option value="Hong Kong">Hong Kong</option>
-		</select>
-	</section> -->
-
 	<section class="stories">
 
 		<div class="search-results">
@@ -37,7 +28,9 @@
 
 	</section>
 
-
+	<script src="scripts/global.js"></script>
+	<script src="scripts/createTemplate.js"></script>
+	<script src="scripts/updateContent.js"></script>
 	<script src="scripts/api.js"></script>
 </body>
