+++ scripts/global.js
@@ -1,0 +1,9 @@
+//defining global variable
+var newsUrl = "https://content.guardianapis.com/search?q="; //the main url
+var newsUrl2 = "&api-key=7b8fdba3-2d40-4d8e-b725-99d86676fbe9&show-fields=all";
+var newsUrlIsrael = newsUrl + 'israel' + newsUrl2;
+var picUrl = "https://countryapi.gear.host/v1/Country/getCountries?pName=";
+var israelUrl = "israel";
+var picUrlIsrael = picUrl + israelUrl;
+var koreaUrl = "Korea+(Republic+of)";
+var ukUrl = "united+kingdom";
